Split user_media module into submodules
