+++ crates/native/src/api/mod.rs
@@ -23,9 +23,9 @@ use crate::{
     frb_generated::{
         FLUTTER_RUST_BRIDGE_CODEGEN_VERSION, RustOpaque, StreamSink,
     },
+    media::TrackOrigin,
     pc::PeerConnectionId,
     renderer::FrameHandler,
-    user_media::TrackOrigin,
 };
 
 /// Custom [`Handler`] for executing Rust code called from Dart.
