--- crates/native/src/user_media.rs
@@ -1,1769 +1,0 @@
-use std::{
-    cmp,
-    collections::{HashMap, HashSet},
-    hash::Hash,
-    mem,
-    sync::{Arc, Mutex, OnceLock, RwLock, Weak},
-};
-
-use anyhow::{Context as _, anyhow, bail};
-use derive_more::with_trait::{AsRef, Display, From, Into};
-use libwebrtc_sys as sys;
-use sys::AudioProcessing;
-use xxhash::xxh3::xxh3_64;
-
-use crate::{
-    PeerConnection, VideoSink, Webrtc, api,
-    api::NoiseSuppressionLevel,
-    devices,
-    frb_generated::StreamSink,
-    next_id,
-    pc::{PeerConnectionId, RtpTransceiver},
-};
-
-impl Webrtc {
-    /// Creates a new [`VideoTrack`]s and/or [`AudioTrack`]s according to the
-    /// provided accepted [`api::MediaStreamConstraints`].
-    pub fn get_media(
-        &mut self,
-        constraints: api::MediaStreamConstraints,
-    ) -> Result<Vec<api::MediaStreamTrack>, api::GetMediaError> {
-        let mut tracks = Vec::new();
-
-        let inner_get_media = || -> Result<(), api::GetMediaError> {
-            if let Some(video) = constraints.video {
-                let src =
-                    self.get_or_create_video_source(&video).map_err(|err| {
-                        api::GetMediaError::Video(err.to_string())
-                    })?;
-                let track = self
-                    .create_video_track(Arc::clone(&src))
-                    .map_err(|err| api::GetMediaError::Video(err.to_string()));
-                if let Err(err) = track {
-                    if Arc::strong_count(&src) == 2 {
-                        self.video_sources.remove(&src.device_id);
-                    }
-                    return Err(err);
-                }
-                tracks.push(track?);
-            }
-
-            if let Some(audio) = constraints.audio {
-                let src = self
-                    .get_or_create_audio_source(&audio)
-                    .map_err(|e| api::GetMediaError::Audio(e.to_string()))?;
-                let track = self
-                    .create_audio_track(Arc::clone(&src))
-                    .map_err(|e| api::GetMediaError::Audio(e.to_string()))?;
-                tracks.push(track);
-            }
-
-            Ok(())
-        };
-
-        if let Err(err) = inner_get_media() {
-            for track in tracks {
-                self.dispose_track(
-                    TrackOrigin::from(
-                        track.peer_id.map(PeerConnectionId::from),
-                    ),
-                    track.id,
-                    track.kind,
-                    true,
-                );
-            }
-
-            Err(err)
-        } else {
-            Ok(tracks)
-        }
-    }
-
-    /// Disposes a [`VideoTrack`] or [`AudioTrack`] by the provided `track_id`.
-    pub fn dispose_track(
-        &mut self,
-        track_origin: TrackOrigin,
-        track_id: String,
-        kind: api::MediaType,
-        notify_on_ended: bool,
-    ) {
-        #[expect(clippy::mutable_key_type, reason = "false positive")]
-        let senders = match kind {
-            api::MediaType::Audio => {
-                if let Some((_, mut track)) = self
-                    .audio_tracks
-                    .remove(&(AudioTrackId::from(track_id), track_origin))
-                {
-                    if let MediaTrackSource::Local(src) = track.source() {
-                        if Arc::strong_count(src) <= 2 {
-                            self.audio_sources.remove(&src.device_id);
-                            self.audio_device_module
-                                .dispose_audio_source(&src.device_id);
-                        }
-                    }
-                    if notify_on_ended {
-                        track.notify_on_ended();
-                    }
-                    mem::take(&mut track.senders)
-                } else {
-                    return;
-                }
-            }
-            api::MediaType::Video => {
-                if let Some((_, mut track)) = self
-                    .video_tracks
-                    .remove(&(VideoTrackId::from(track_id), track_origin))
-                {
-                    for id in track.kind.sinks.clone() {
-                        if let Some(sink) = self.video_sinks.remove(&id) {
-                            track.remove_video_sink(sink);
-                        }
-                    }
-                    if let MediaTrackSource::Local(src) = track.source() {
-                        if Arc::strong_count(src) == 2 {
-                            self.video_sources.remove(&src.device_id);
-                        }
-                    }
-                    if notify_on_ended {
-                        track.notify_on_ended();
-                    }
-                    track.senders.clone()
-                } else {
-                    return;
-                }
-            }
-        };
-
-        #[expect(clippy::iter_over_hash_type, reason = "doesn't matter")]
-        for (peer, tranceivers) in senders {
-            for transceiver in tranceivers {
-                if let Err(e) =
-                    self.sender_replace_track(&peer, &transceiver, None)
-                {
-                    log::error!("Failed to remove track for the sender: {e}");
-                }
-            }
-        }
-    }
-
-    /// Creates a new [`VideoTrack`] from the given [`VideoSource`].
-    fn create_video_track(
-        &self,
-        source: Arc<VideoSource>,
-    ) -> anyhow::Result<api::MediaStreamTrack> {
-        let track =
-            VideoTrack::create_local(&self.peer_connection_factory, source)?;
-
-        let api_track = api::MediaStreamTrack::from(&track);
-
-        self.video_tracks.insert((track.id(), TrackOrigin::Local), track);
-
-        Ok(api_track)
-    }
-
-    /// Creates a new [`VideoSource`] based on the given [`VideoConstraints`].
-    fn get_or_create_video_source(
-        &mut self,
-        caps: &api::VideoConstraints,
-    ) -> anyhow::Result<Arc<VideoSource>> {
-        let (source, device_id) = if caps.is_display {
-            let device_id = if let Some(device_id) = caps.device_id.clone() {
-                sys::screen_capture_sources()
-                    .into_iter()
-                    .find(|d| d.id().to_string() == device_id)
-                    .ok_or_else(|| {
-                        anyhow!(
-                            "Cannot find video display with the specified ID: \
-                             {device_id}",
-                        )
-                    })?;
-                VideoDeviceId(device_id)
-            } else {
-                let displays = devices::enumerate_displays();
-                // No device ID is provided, so just pick the first available
-                // device.
-                if displays.is_empty() {
-                    bail!("Cannot find any available video input displays");
-                }
-
-                VideoDeviceId(displays[0].device_id.clone())
-            };
-            if let Some(src) = self.video_sources.get(&device_id) {
-                return Ok(Arc::clone(src));
-            }
-
-            (
-                VideoSource::new_display_source(
-                    &mut self.worker_thread,
-                    &mut self.signaling_thread,
-                    caps,
-                    device_id.clone(),
-                )?,
-                device_id,
-            )
-        } else {
-            let (index, device_id) =
-                if let Some(device_id) = caps.device_id.clone() {
-                    let device_id = VideoDeviceId(device_id);
-                    if let Some(index) =
-                        self.get_index_of_video_device(&device_id)?
-                    {
-                        (index, device_id)
-                    } else {
-                        bail!(
-                            "Cannot find video device with the specified ID: \
-                             {device_id}",
-                        );
-                    }
-                } else {
-                    // No device ID is provided, so just pick the first
-                    // available device.
-                    if self.video_device_info.number_of_devices() < 1 {
-                        bail!("Cannot find any available video input devices");
-                    }
-
-                    let device_id =
-                        VideoDeviceId(self.video_device_info.device_name(0)?.1);
-                    (0, device_id)
-                };
-            if let Some(src) = self.video_sources.get(&device_id) {
-                return Ok(Arc::clone(src));
-            }
-
-            (
-                VideoSource::new_device_source(
-                    &mut self.worker_thread,
-                    &mut self.signaling_thread,
-                    caps,
-                    index,
-                    device_id.clone(),
-                )?,
-                device_id,
-            )
-        };
-
-        let source = self
-            .video_sources
-            .entry(device_id)
-            .or_insert_with(|| Arc::new(source));
-
-        Ok(Arc::clone(source))
-    }
-
-    /// Creates a new [`AudioTrack`] from the given
-    /// [`sys::AudioSourceInterface`].
-    fn create_audio_track(
-        &self,
-        source: Arc<AudioSource>,
-    ) -> anyhow::Result<api::MediaStreamTrack> {
-        let track = AudioTrack::new(
-            &self.peer_connection_factory,
-            source,
-            TrackOrigin::Local,
-        )?;
-
-        let api_track = api::MediaStreamTrack::from(&track);
-
-        self.audio_tracks.insert((track.id(), TrackOrigin::Local), track);
-
-        Ok(api_track)
-    }
-
-    /// Creates a new [`sys::AudioSourceInterface`] based on the given
-    /// [`AudioConstraints`].
-    fn get_or_create_audio_source(
-        &mut self,
-        caps: &api::AudioConstraints,
-    ) -> anyhow::Result<Arc<AudioSource>> {
-        let device_id = if let Some(device_id) = caps.device_id.clone() {
-            AudioDeviceId(device_id)
-        } else {
-            // `AudioDeviceModule` is not capturing anything at the moment,
-            // so we will use first available device (with `0` index).
-            if self.audio_device_module.recording_devices() < 1 {
-                bail!("Cannot find any available audio input device");
-            }
-
-            AudioDeviceId(self.audio_device_module.recording_device_name(0)?.1)
-        };
-
-        let Some(device_index) =
-            self.get_index_of_audio_recording_device(&device_id)?
-        else {
-            bail!(
-                "Cannot find audio device with the specified ID `{device_id}`",
-            );
-        };
-
-        let src = if let Some(src) = self.audio_sources.get(&device_id) {
-            src.update_audio_processing(&caps.processing);
-            Arc::clone(src)
-        } else {
-            let processing =
-                sys::AudioProcessing::new((&caps.processing).into())?;
-
-            let src = Arc::new(AudioSource::new(
-                device_id.clone(),
-                self.audio_device_module
-                    .create_audio_source(device_index, &processing)?,
-                processing,
-            ));
-            self.audio_sources.insert(device_id, Arc::clone(&src));
-
-            src
-        };
-
-        Ok(src)
-    }
-
-    /// Returns the [readyState][0] property of the media track by its ID and
-    /// media type.
-    ///
-    /// [0]: https://w3.org/TR/mediacapture-streams#dfn-readystate
-    #[must_use]
-    pub fn track_state(
-        &self,
-        id: String,
-        track_origin: TrackOrigin,
-        kind: api::MediaType,
-    ) -> api::TrackState {
-        match kind {
-            api::MediaType::Audio => {
-                let id = AudioTrackId::from(id);
-                self.audio_tracks
-                    .get(&(id, track_origin))
-                    .map_or_else(|| api::TrackState::Ended, |t| t.state())
-            }
-            api::MediaType::Video => {
-                let id = VideoTrackId::from(id);
-                self.video_tracks
-                    .get(&(id, track_origin))
-                    .map_or_else(|| api::TrackState::Ended, |t| t.state())
-            }
-        }
-    }
-
-    /// Returns the [width] property of the media track by its ID and origin.
-    ///
-    /// Blocks until the [width] is initialized.
-    ///
-    /// [width]: https://w3.org/TR/mediacapture-streams#dfn-width
-    #[must_use]
-    pub fn track_width(
-        &self,
-        id: String,
-        track_origin: TrackOrigin,
-    ) -> Option<i32> {
-        let id = VideoTrackId::from(id);
-
-        self.video_tracks
-            .get(&(id, track_origin))
-            .map(|t| *t.kind.width.wait().read().unwrap())
-    }
-
-    /// Returns the [height] property of the media track by its ID and origin.
-    ///
-    /// Blocks until the [height] is initialized.
-    ///
-    /// [height]: https://w3.org/TR/mediacapture-streams#dfn-height
-    #[must_use]
-    pub fn track_height(
-        &self,
-        id: String,
-        track_origin: TrackOrigin,
-    ) -> Option<i32> {
-        let id = VideoTrackId::from(id);
-
-        self.video_tracks
-            .get(&(id, track_origin))
-            .map(|t| *t.kind.height.wait().read().unwrap())
-    }
-
-    /// Changes the [enabled][1] property of the media track by its ID.
-    ///
-    /// [1]: https://w3.org/TR/mediacapture-streams#track-enabled
-    pub fn set_track_enabled(
-        &self,
-        id: String,
-        track_origin: TrackOrigin,
-        kind: api::MediaType,
-        enabled: bool,
-    ) {
-        match kind {
-            api::MediaType::Audio => {
-                let id = AudioTrackId::from(id);
-                let track = self.audio_tracks.get(&(id, track_origin));
-                if let Some(track) = track {
-                    track.set_enabled(enabled);
-                }
-            }
-            api::MediaType::Video => {
-                let id = VideoTrackId::from(id);
-                let track = self.video_tracks.get(&(id, track_origin));
-                if let Some(track) = track {
-                    track.set_enabled(enabled);
-                }
-            }
-        }
-    }
-
-    /// Clones the specified [`api::MediaStreamTrack`].
-    #[must_use]
-    pub fn clone_track(
-        &self,
-        id: String,
-        track_origin: TrackOrigin,
-        kind: api::MediaType,
-    ) -> Option<api::MediaStreamTrack> {
-        match kind {
-            api::MediaType::Audio => {
-                let id = AudioTrackId::from(id);
-                let source = self.audio_tracks.get(&(id, track_origin)).map(
-                    |track| match track.source() {
-                        MediaTrackSource::Local(source) => {
-                            MediaTrackSource::Local(Arc::clone(source))
-                        }
-                        MediaTrackSource::Remote { mid, peer } => {
-                            MediaTrackSource::Remote {
-                                mid: mid.clone(),
-                                peer: Weak::clone(peer),
-                            }
-                        }
-                    },
-                )?;
-
-                match source {
-                    MediaTrackSource::Local(source) => Some(
-                        self.create_audio_track(Arc::clone(&source)).ok()?,
-                    ),
-                    MediaTrackSource::Remote { mid, peer } => {
-                        let peer = peer.upgrade()?;
-                        let mut transceivers = peer.get_transceivers();
-
-                        transceivers.retain(|transceiver| {
-                            transceiver.mid().unwrap() == mid
-                        });
-
-                        if transceivers.is_empty() {
-                            return None;
-                        }
-
-                        let track = AudioTrack::wrap_remote(
-                            transceivers.first().unwrap(),
-                            &peer,
-                        );
-
-                        Some(api::MediaStreamTrack::from(&track))
-                    }
-                }
-            }
-            api::MediaType::Video => {
-                let id = VideoTrackId::from(id);
-                let source = self.video_tracks.get(&(id, track_origin)).map(
-                    |track| match track.source() {
-                        MediaTrackSource::Local(source) => {
-                            MediaTrackSource::Local(Arc::clone(source))
-                        }
-                        MediaTrackSource::Remote { mid, peer } => {
-                            MediaTrackSource::Remote {
-                                mid: mid.clone(),
-                                peer: Weak::clone(peer),
-                            }
-                        }
-                    },
-                )?;
-
-                match source {
-                    MediaTrackSource::Local(source) => {
-                        Some(self.create_video_track(source).ok()?)
-                    }
-                    MediaTrackSource::Remote { mid, peer } => {
-                        let peer = peer.upgrade()?;
-                        let mut transceivers = peer.get_transceivers();
-                        transceivers.retain(|transceiver| {
-                            transceiver.mid().unwrap() == mid
-                        });
-
-                        if transceivers.is_empty() {
-                            return None;
-                        }
-
-                        let track = VideoTrack::wrap_remote(
-                            transceivers.first().unwrap(),
-                            &peer,
-                        );
-
-                        Some(api::MediaStreamTrack::from(&track))
-                    }
-                }
-            }
-        }
-    }
-
-    /// Enables or disables audio level observing of the [`AudioTrack`] with the
-    /// provided `id`.
-    ///
-    /// # Warning
-    ///
-    /// Returns error message if cannot find any [`AudioTrack`] by the provided
-    /// `id`.
-    pub fn set_audio_level_observer_enabled(
-        &self,
-        id: String,
-        track_origin: TrackOrigin,
-        enabled: bool,
-    ) {
-        let id = AudioTrackId::from(id);
-        let track = self.audio_tracks.get_mut(&(id, track_origin));
-        if let Some(mut track) = track {
-            if enabled {
-                track.subscribe_to_audio_level();
-            } else {
-                track.unsubscribe_from_audio_level();
-            }
-        }
-    }
-
-    /// Registers an events observer for an [`AudioTrack`] or a [`VideoTrack`].
-    ///
-    /// # Warning
-    ///
-    /// Returns error message if cannot find any [`AudioTrack`] or
-    /// [`VideoTrack`] by the specified `id`.
-    pub fn register_track_observer(
-        &self,
-        id: String,
-        track_origin: TrackOrigin,
-        kind: api::MediaType,
-        track_events_tx: StreamSink<api::TrackEvent>,
-    ) {
-        let mut obs = sys::TrackEventObserver::new(Box::new(
-            TrackEventHandler::new(track_events_tx.clone()),
-        ));
-        match kind {
-            api::MediaType::Audio => {
-                let id = AudioTrackId::from(id);
-                let track = self.audio_tracks.get_mut(&(id, track_origin));
-
-                if let Some(mut track) = track {
-                    obs.set_audio_track(&track.kind.inner);
-                    track.set_track_events_tx(track_events_tx);
-                    track.kind.inner.register_observer(obs);
-                }
-            }
-            api::MediaType::Video => {
-                let id = VideoTrackId::from(id);
-                let track = self.video_tracks.get_mut(&(id, track_origin));
-
-                if let Some(mut track) = track {
-                    obs.set_video_track(&track.kind.inner);
-                    track.set_track_events_tx(track_events_tx);
-                    track.kind.inner.register_observer(obs);
-                }
-            }
-        }
-    }
-
-    /// Applies the provided [`api::AudioProcessingConstraints`] to the
-    /// [`sys::AudioSourceInterface`] of the referred local [`AudioTrack`].
-    ///
-    /// # Errors
-    ///
-    /// If the provided [`AudioTrackId`] refers to a remote [`AudioTrack`].
-    pub fn apply_audio_processing_config(
-        &self,
-        id: String,
-        conf: &api::AudioProcessingConstraints,
-    ) -> anyhow::Result<()> {
-        let id = AudioTrackId::from(id);
-        let Some(track) = self.audio_tracks.get_mut(&(id, TrackOrigin::Local))
-        else {
-            return Ok(());
-        };
-
-        let MediaTrackSource::Local(src) = track.source() else {
-            bail!("Cannot change audio processing of remote `AudioTrack`");
-        };
-
-        src.update_audio_processing(conf);
-
-        Ok(())
-    }
-
-    /// Returns the [`api::AudioProcessingConstraints`] of the
-    /// [`sys::AudioSourceInterface`] of the referred local [`AudioTrack`].
-    ///
-    /// # Errors
-    ///
-    /// If the provided [`AudioTrackId`] refers to a remote [`AudioTrack`].
-    pub fn get_audio_processing_config(
-        &self,
-        id: String,
-    ) -> anyhow::Result<api::AudioProcessingConfig> {
-        let id = AudioTrackId::from(id);
-        let Some(track) = self.audio_tracks.get_mut(&(id, TrackOrigin::Local))
-        else {
-            // This means that the `AudioTrack` was already removed, so to omit
-            // erroring something should be returned. It shouldn't really matter
-            // what is returned, so default values are just okay.
-            return Ok(api::AudioProcessingConfig {
-                auto_gain_control: true,
-                high_pass_filter: true,
-                noise_suppression: true,
-                noise_suppression_level: NoiseSuppressionLevel::VeryHigh,
-                echo_cancellation: true,
-            });
-        };
-
-        let MediaTrackSource::Local(src) = track.source() else {
-            bail!("Cannot get audio processing of remote `AudioTrack`");
-        };
-
-        let mut conf = src.ap.config();
-
-        Ok(api::AudioProcessingConfig {
-            auto_gain_control: conf.get_gain_controller_enabled(),
-            high_pass_filter: conf.get_high_pass_filter_enabled(),
-            noise_suppression: conf.get_noise_suppression_enabled(),
-            noise_suppression_level: conf.get_noise_suppression_level().into(),
-            echo_cancellation: conf.get_echo_cancellation_enabled(),
-        })
-    }
-}
-
-/// ID of a [MediaStream].
-///
-/// [MediaStream]: https://w3.org/TR/mediacapture-streams#dom-mediastream
-#[derive(Clone, Copy, Debug, Display, Eq, From, Hash, PartialEq)]
-pub struct MediaStreamId(u64);
-
-/// ID of an video input device that provides data to some [`VideoSource`].
-#[derive(AsRef, Clone, Debug, Display, Eq, From, Hash, Into, PartialEq)]
-#[as_ref(forward)]
-pub struct VideoDeviceId(String);
-
-/// ID of an `AudioDevice`.
-#[derive(
-    AsRef, Clone, Debug, Default, Display, Eq, From, Hash, Into, PartialEq,
-)]
-#[as_ref(forward)]
-pub struct AudioDeviceId(String);
-
-/// ID of a [`VideoTrack`].
-#[derive(Clone, Debug, Display, From, Eq, Hash, Into, PartialEq)]
-pub struct VideoTrackId(String);
-
-/// ID of an [`AudioTrack`].
-#[derive(Clone, Debug, Display, From, Eq, Hash, Into, PartialEq)]
-pub struct AudioTrackId(String);
-
-/// Label identifying a video track source.
-#[derive(AsRef, Clone, Debug, Default, Display, Eq, From, Hash, PartialEq)]
-#[as_ref(forward)]
-#[from(forward)]
-pub struct VideoLabel(String);
-
-/// Label identifying an audio track source.
-#[derive(AsRef, Clone, Debug, Default, Display, Eq, From, Hash, PartialEq)]
-#[as_ref(forward)]
-#[from(forward)]
-pub struct AudioLabel(String);
-
-/// [`sys::VideoDeviceInfo`] wrapper.
-pub struct VideoDeviceInfo(sys::VideoDeviceInfo);
-
-impl VideoDeviceInfo {
-    /// Creates a new [`VideoDeviceInfo`].
-    ///
-    /// # Errors
-    ///
-    /// If [`sys::VideoDeviceInfo::create()`] returns error.
-    pub fn new() -> anyhow::Result<Self> {
-        Ok(Self(sys::VideoDeviceInfo::create()?))
-    }
-
-    /// Returns count of a video recording devices.
-    pub fn number_of_devices(&mut self) -> u32 {
-        if api::is_fake_media() { 1 } else { self.0.number_of_devices() }
-    }
-
-    /// Returns the `(label, id)` tuple for the given video device `index`.
-    ///
-    /// # Errors
-    ///
-    /// If [`sys::VideoDeviceInfo::device_name()`] call returns error.
-    pub fn device_name(
-        &mut self,
-        index: u32,
-    ) -> anyhow::Result<(String, String)> {
-        if api::is_fake_media() {
-            Ok((String::from("fake camera"), String::from("fake camera id")))
-        } else {
-            self.0.device_name(index)
-        }
-    }
-}
-
-/// [`sys::AudioDeviceModule`] wrapper tracking the currently used audio input
-/// device.
-#[derive(AsRef)]
-pub struct AudioDeviceModule {
-    /// [`sys::AudioDeviceModule`] backing this [`AudioDeviceModule`].
-    #[as_ref]
-    inner: sys::AudioDeviceModule,
-}
-
-impl AudioDeviceModule {
-    /// Creates a new [`AudioDeviceModule`] according to the passed
-    /// [`sys::AudioLayer`].
-    ///
-    /// # Errors
-    ///
-    /// If could not find any available recording device.
-    pub fn new(
-        worker_thread: &mut sys::Thread,
-        audio_layer: sys::AudioLayer,
-        task_queue_factory: &mut sys::TaskQueueFactory,
-    ) -> anyhow::Result<Self> {
-        let inner = sys::AudioDeviceModule::create_proxy(
-            worker_thread,
-            audio_layer,
-            task_queue_factory,
-        )?;
-        inner.init()?;
-
-        Ok(Self { inner })
-    }
-
-    /// Returns the `(label, id)` tuple for the given audio playout device
-    /// `index`.
-    ///
-    /// # Errors
-    ///
-    /// If [`sys::AudioDeviceModule::playout_device_name()`] call fails.
-    pub fn playout_device_name(
-        &self,
-        index: i16,
-    ) -> anyhow::Result<(String, String)> {
-        if api::is_fake_media() {
-            return Ok((
-                String::from("fake headset"),
-                String::from("fake headset id"),
-            ));
-        }
-
-        let (label, mut device_id) = self.inner.playout_device_name(index)?;
-
-        if device_id.is_empty() {
-            let hash = xxh3_64(
-                [label.as_bytes(), &[api::MediaDeviceKind::AudioOutput as u8]]
-                    .concat()
-                    .as_slice(),
-            );
-            device_id = hash.to_string();
-        }
-
-        Ok((label, device_id))
-    }
-
-    /// Returns the `(label, id)` tuple for the given audio recording device
-    /// `index`.
-    ///
-    /// # Errors
-    ///
-    /// If [`sys::AudioDeviceModule::recording_device_name()`] call fails.
-    pub fn recording_device_name(
-        &self,
-        index: i16,
-    ) -> anyhow::Result<(String, String)> {
-        if api::is_fake_media() {
-            return Ok((String::from("fake mic"), String::from("fake mic id")));
-        }
-
-        let (label, mut device_id) = self.inner.recording_device_name(index)?;
-
-        if device_id.is_empty() {
-            let hash = xxh3_64(
-                [label.as_bytes(), &[api::MediaDeviceKind::AudioOutput as u8]]
-                    .concat()
-                    .as_slice(),
-            );
-            device_id = hash.to_string();
-        }
-
-        Ok((label, device_id))
-    }
-
-    /// Returns count of available audio playout devices.
-    ///
-    /// # Errors
-    ///
-    /// If [`sys::AudioDeviceModule::playout_devices()`] call fails.
-    #[must_use]
-    pub fn playout_devices(&self) -> u32 {
-        self.inner.playout_devices()
-    }
-
-    /// Returns count of available audio recording devices.
-    ///
-    /// # Errors
-    ///
-    /// If [`sys::AudioDeviceModule::recording_devices()`] call fails.
-    #[must_use]
-    pub fn recording_devices(&self) -> u32 {
-        if api::is_fake_media() { 1 } else { self.inner.recording_devices() }
-    }
-
-    /// Creates a new [`sys::AudioSourceInterface`] based on the provided
-    /// `device_index`.
-    ///
-    /// # Errors
-    ///
-    /// If [`sys::AudioDeviceModule::recording_devices()`] call fails.
-    pub fn create_audio_source(
-        &mut self,
-        device_index: u16,
-        ap: &AudioProcessing,
-    ) -> anyhow::Result<sys::AudioSourceInterface> {
-        if api::is_fake_media() {
-            self.inner.create_fake_audio_source()
-        } else {
-            self.inner.create_audio_source(device_index, ap)
-        }
-    }
-
-    /// Disposes a [`sys::AudioSourceInterface`] by the provided
-    /// [`AudioDeviceId`].
-    pub fn dispose_audio_source(&mut self, device_id: &AudioDeviceId) {
-        self.inner.dispose_audio_source(device_id.to_string());
-    }
-
-    /// Sets the microphone system volume according to the given level in
-    /// percents.
-    ///
-    /// # Errors
-    ///
-    /// Errors if any of the following calls fail:
-    ///     - [`sys::AudioDeviceModule::microphone_volume_is_available()`];
-    ///     - [`sys::AudioDeviceModule::min_microphone_volume()`];
-    ///     - [`sys::AudioDeviceModule::max_microphone_volume()`];
-    ///     - [`sys::AudioDeviceModule::set_microphone_volume()`].
-    pub fn set_microphone_volume(&self, mut level: u8) -> anyhow::Result<()> {
-        if !self.microphone_volume_is_available()? {
-            bail!("The microphone volume is unavailable.")
-        }
-
-        if level > 100 {
-            level = 100;
-        }
-
-        let min_volume = self.inner.min_microphone_volume()?;
-        let max_volume = self.inner.max_microphone_volume()?;
-
-        let volume = f64::from(max_volume - min_volume)
-            .mul_add(f64::from(level) / 100.0, f64::from(min_volume));
-
-        #[expect( // intentional
-            clippy::cast_possible_truncation,
-            clippy::cast_sign_loss,
-            reason = "size fits and non-negative"
-        )]
-        self.inner.set_microphone_volume(volume as u32)
-    }
-
-    /// Indicates if the microphone is available to set volume.
-    ///
-    /// # Errors
-    ///
-    /// If [`sys::AudioDeviceModule::microphone_volume_is_available()`] call
-    /// fails.
-    pub fn microphone_volume_is_available(&self) -> anyhow::Result<bool> {
-        Ok(self.inner.microphone_volume_is_available().unwrap_or(false))
-    }
-
-    /// Returns the current level of the microphone volume in percents.
-    ///
-    /// # Errors
-    ///
-    /// If fails on:
-    ///     - [`sys::AudioDeviceModule::microphone_volume()`] call
-    ///     - [`sys::AudioDeviceModule::min_microphone_volume()`] call
-    ///     - [`sys::AudioDeviceModule::max_microphone_volume()`] call
-    pub fn microphone_volume(&self) -> anyhow::Result<u32> {
-        let volume = self.inner.microphone_volume()?;
-        let min_volume = self.inner.min_microphone_volume()?;
-        let max_volume = self.inner.max_microphone_volume()?;
-
-        #[expect( // intentional
-            clippy::cast_possible_truncation,
-            clippy::cast_sign_loss,
-            reason = "size fits and non-negative"
-        )]
-        let level = (f64::from(volume - min_volume)
-            / f64::from(max_volume - min_volume)
-            * 100.0) as u32;
-
-        Ok(level)
-    }
-
-    /// Changes the playout device for this [`AudioDeviceModule`].
-    ///
-    /// # Errors
-    ///
-    /// If [`sys::AudioDeviceModule::set_playout_device()`] call fails.
-    pub fn set_playout_device(&self, index: u16) -> anyhow::Result<()> {
-        self.inner.set_playout_device(index)?;
-
-        Ok(())
-    }
-
-    /// Stops playout of audio on this [`AudioDeviceModule`].
-    ///
-    /// # Errors
-    ///
-    /// If [`sys::AudioDeviceModule::stop_playout()`] call fails.
-    pub fn stop_playout(&self) -> anyhow::Result<()> {
-        self.inner.stop_playout()
-    }
-
-    /// Indicates whether stereo is available in this playout
-    /// [`AudioDeviceModule`].
-    ///
-    /// # Errors
-    ///
-    /// If [`sys::AudioDeviceModule::stereo_playout_is_available()`] call fails.
-    pub fn stereo_playout_is_available(&self) -> anyhow::Result<bool> {
-        self.inner.stereo_playout_is_available()
-    }
-
-    /// Initializes this playout [`AudioDeviceModule`].
-    ///
-    /// # Errors
-    ///
-    /// If [`sys::AudioDeviceModule::init_playout()`] call fails.
-    pub fn init_playout(&self) -> anyhow::Result<()> {
-        self.inner.init_playout()
-    }
-
-    /// Starts playout of audio on this [`AudioDeviceModule`].
-    ///
-    /// # Errors
-    ///
-    /// If [`sys::AudioDeviceModule::start_playout()`] call fails.
-    pub fn start_playout(&self) -> anyhow::Result<()> {
-        self.inner.start_playout()
-    }
-}
-
-/// Indicates whether some track is a local track obtained via
-/// [getUserMedia()][1]/[getDisplayMedia()][2] call or a remote received in a
-/// [ontrack][3] callback.
-///
-/// [1]: https://w3.org/TR/mediacapture-streams#dom-mediadevices-getusermedia
-/// [2]: https://w3.org/TR/screen-capture/#dom-mediadevices-getdisplaymedia
-/// [3]: https://w3.org/TR/webrtc/#dom-rtcpeerconnection-ontrack
-#[derive(Clone, Copy, Debug, Eq, From, Hash, PartialEq)]
-pub enum TrackOrigin {
-    Local,
-    Remote(PeerConnectionId),
-}
-
-impl From<Option<PeerConnectionId>> for TrackOrigin {
-    fn from(value: Option<PeerConnectionId>) -> Self {
-        value.map_or(Self::Local, Self::Remote)
-    }
-}
-
-/// Possible kinds of media track's source.
-pub enum MediaTrackSource<T> {
-    Local(Arc<T>),
-    Remote { mid: String, peer: Weak<PeerConnection> },
-}
-
-mod kind {
-    use std::sync::{Arc, OnceLock, RwLock};
-
-    use derive_more::AsRef;
-    use libwebrtc_sys as sys;
-
-    use crate::{
-        AudioSource, AudioTrackId, VideoSink, VideoSinkId, VideoSource,
-        VideoTrackId, api,
-        frb_generated::StreamSink,
-        user_media::{
-            AudioLevelObserverId, AudioSourceAudioLevelHandler,
-            MediaTrackSource,
-        },
-    };
-
-    /// Representation of a [`sys::VideoTrackInterface`].
-    #[derive(AsRef)]
-    pub struct Video {
-        /// ID of this [`Video`].
-        pub id: VideoTrackId,
-
-        /// [`MediaTrackSource`] that is used by this [`Video`].
-        pub source: MediaTrackSource<VideoSource>,
-
-        /// Underlying [`sys::VideoTrackInterface`].
-        #[as_ref]
-        pub inner: sys::VideoTrackInterface,
-
-        /// List of the [`VideoSink`]s attached to this [`VideoTrack`].
-        pub sinks: Vec<VideoSinkId>,
-
-        /// Tracks changes in video `height` and `width`.
-        pub sink: Option<VideoSink>,
-
-        /// Video width.
-        pub width: Arc<OnceLock<RwLock<i32>>>,
-
-        /// Video height.
-        pub height: Arc<OnceLock<RwLock<i32>>>,
-    }
-
-    impl Video {
-        /// Adds the provided [`VideoSink`] to this [`Video`].
-        pub fn add_video_sink(&mut self, video_sink: &mut VideoSink) {
-            self.inner.add_or_update_sink(video_sink.as_mut());
-            self.sinks.push(video_sink.id());
-        }
-
-        /// Detaches the provided [`VideoSink`] from this [`Video`].
-        pub fn remove_video_sink(&mut self, mut video_sink: VideoSink) {
-            self.sinks.retain(|&sink| sink != video_sink.id());
-            self.inner.remove_sink(video_sink.as_mut());
-        }
-    }
-
-    impl Drop for Video {
-        fn drop(&mut self) {
-            let sink = self.sink.take().unwrap();
-            self.remove_video_sink(sink);
-        }
-    }
-
-    /// Representation of a [`sys::AudioTrackInterface`].
-    #[derive(AsRef)]
-    pub struct Audio {
-        /// ID of this [`Audio`].
-        pub id: AudioTrackId,
-
-        /// [`AudioSource`] that is used by this [`Audio`].
-        pub source: MediaTrackSource<AudioSource>,
-
-        /// Underlying [`sys::AudioTrackInterface`].
-        #[as_ref]
-        pub inner: sys::AudioTrackInterface,
-
-        /// [`AudioLevelObserverId`] related to this [`Audio`].
-        ///
-        /// This ID can be used when this [`Audio`] needs to dispose its
-        /// observer.
-        pub volume_observer_id: Option<AudioLevelObserverId>,
-    }
-
-    impl Audio {
-        /// Subscribes this [`Audio`] to audio level updates.
-        ///
-        /// Volume updates will be passed to the `stream_sink` of this
-        /// [`Audio`].
-        pub fn subscribe_to_audio_level(
-            &mut self,
-            track_events_tx: Option<StreamSink<api::TrackEvent>>,
-        ) {
-            if let Some(sink) = track_events_tx {
-                match &self.source {
-                    MediaTrackSource::Local(src) => {
-                        let observer = src.subscribe_on_audio_level(
-                            AudioSourceAudioLevelHandler(sink),
-                        );
-                        self.volume_observer_id = Some(observer);
-                    }
-                    MediaTrackSource::Remote { mid: _, peer: _ } => (),
-                }
-            }
-        }
-
-        /// Unsubscribes this [`Audio`] from audio level updates.
-        pub fn unsubscribe_from_audio_level(&self) {
-            match &self.source {
-                MediaTrackSource::Local(src) => {
-                    if let Some(id) = self.volume_observer_id {
-                        src.unsubscribe_audio_level(id);
-                    }
-                }
-                MediaTrackSource::Remote { mid: _, peer: _ } => (),
-            }
-        }
-    }
-
-    impl Drop for Audio {
-        fn drop(&mut self) {
-            self.unsubscribe_from_audio_level();
-        }
-    }
-}
-
-/// Representation of a generic track interface.
-pub struct Track<T> {
-    /// Indicator whether this is a remote or a local track.
-    origin: TrackOrigin,
-
-    /// [`TrackInterface`] of this [`Track`].
-    kind: T,
-
-    /// `StreamSink` which can be used by this [`Track`] to emit
-    /// [`api::TrackEvent`]s to Flutter side.
-    events_tx: Option<StreamSink<api::TrackEvent>>,
-
-    /// Peers and transceivers sending this [`Track`].
-    senders: HashMap<Arc<PeerConnection>, HashSet<Arc<RtpTransceiver>>>,
-}
-
-impl<T> Track<T> {
-    /// Emits [`api::TrackEvent::Ended`] to the Flutter side.
-    pub fn notify_on_ended(&mut self) {
-        if let Some(sink) = self.events_tx.take() {
-            _ = sink.add(api::TrackEvent::Ended);
-        }
-    }
-
-    /// Sets the provided `StreamSink` for this [`Track`] to use for
-    /// [`api::TrackEvent`]s emitting.
-    pub fn set_track_events_tx(&mut self, sink: StreamSink<api::TrackEvent>) {
-        drop(self.events_tx.replace(sink));
-    }
-
-    /// Adds transceiver to senders of this [`Track`].
-    pub fn add_transceiver(
-        &mut self,
-        peer: Arc<PeerConnection>,
-        transceiver: Arc<RtpTransceiver>,
-    ) {
-        self.senders.entry(peer).or_default().insert(transceiver);
-    }
-
-    /// Removes transceiver from senders of this [`Track`].
-    pub fn remove_transceiver(
-        &mut self,
-        peer: &Arc<PeerConnection>,
-        transceiver: &Arc<RtpTransceiver>,
-    ) {
-        if let Some(transceivers) = self.senders.get_mut(peer) {
-            transceivers.retain(|current| current != transceiver);
-
-            if !transceivers.is_empty() {
-                return;
-            }
-        }
-
-        self.senders.remove(peer);
-    }
-
-    /// Removes peer and its transceivers from senders of this [`Track`].
-    pub fn remove_peer(&mut self, peer: &Arc<PeerConnection>) {
-        self.senders.remove(peer);
-    }
-}
-
-/// Representation of a `Video` track interface.
-pub type VideoTrack = Track<kind::Video>;
-
-impl AsRef<sys::VideoTrackInterface> for VideoTrack {
-    fn as_ref(&self) -> &sys::VideoTrackInterface {
-        self.kind.as_ref()
-    }
-}
-
-/// Tracks changes in video `height` and `width`.
-struct VideoFormatSink {
-    /// Video width.
-    width: Arc<OnceLock<RwLock<i32>>>,
-
-    /// Video height.
-    height: Arc<OnceLock<RwLock<i32>>>,
-}
-
-impl sys::OnFrameCallback for VideoFormatSink {
-    fn on_frame(&mut self, frame: cxx::UniquePtr<sys::VideoFrame>) {
-        if self.width.get().is_none() {
-            self.width.set(RwLock::from(frame.width())).unwrap();
-            self.height.set(RwLock::from(frame.height())).unwrap();
-        } else {
-            *self.width.get().unwrap().write().unwrap() = frame.width();
-            *self.height.get().unwrap().write().unwrap() = frame.height();
-        }
-    }
-}
-
-impl VideoTrack {
-    /// Returns ID of this [`VideoTrack`].
-    #[must_use]
-    pub fn id(&self) -> VideoTrackId {
-        self.kind.id.clone()
-    }
-
-    /// Returns [`VideoSource`] that is used by this [`VideoTrack`].
-    #[must_use]
-    pub const fn source(&self) -> &MediaTrackSource<VideoSource> {
-        &self.kind.source
-    }
-
-    /// Returns the [readyState][0] property of the
-    /// [`sys::VideoTrackInterface`].
-    ///
-    /// [0]: https://w3.org/TR/mediacapture-streams#dfn-readystate
-    fn state(&self) -> api::TrackState {
-        self.kind.inner.state().into()
-    }
-
-    /// Changes the [enabled][1] property ofthe [`sys::VideoTrackInterface`].
-    ///
-    /// [1]: https://w3.org/TR/mediacapture-streams#track-enabled
-    fn set_enabled(&self, enabled: bool) {
-        self.kind.inner.set_enabled(enabled);
-    }
-
-    /// Creates a new [`VideoTrack`].
-    fn create_local(
-        pc: &sys::PeerConnectionFactoryInterface,
-        src: Arc<VideoSource>,
-    ) -> anyhow::Result<Self> {
-        let id = VideoTrackId(next_id().to_string());
-        let track_origin = TrackOrigin::Local;
-
-        let width = Arc::new(OnceLock::new());
-        let height = Arc::new(OnceLock::new());
-        let mut sink = VideoSink::new(
-            i64::from(next_id()),
-            sys::VideoSinkInterface::create_forwarding(Box::new(
-                VideoFormatSink {
-                    width: Arc::clone(&width),
-                    height: Arc::clone(&height),
-                },
-            )),
-            id.clone(),
-            track_origin,
-        );
-
-        let mut res = Self {
-            kind: kind::Video {
-                id: id.clone(),
-                inner: pc.create_video_track(id.into(), &src.inner)?,
-                sinks: Vec::new(),
-                width,
-                height,
-                sink: None,
-                source: MediaTrackSource::Local(src),
-            },
-            senders: HashMap::new(),
-            events_tx: None,
-            origin: track_origin,
-        };
-
-        res.add_video_sink(&mut sink);
-        res.kind.sink = Some(sink);
-
-        Ok(res)
-    }
-
-    /// Wraps the track of the `transceiver.receiver.track()` into a
-    /// [`VideoTrack`].
-    pub(crate) fn wrap_remote(
-        transceiver: &sys::RtpTransceiverInterface,
-        peer: &Arc<PeerConnection>,
-    ) -> Self {
-        let receiver = transceiver.receiver();
-        let track = receiver.track();
-        let track_origin = TrackOrigin::Remote(peer.id());
-
-        let width = Arc::new(OnceLock::new());
-        width.set(RwLock::from(0)).unwrap();
-        let height = Arc::new(OnceLock::new());
-        height.set(RwLock::from(0)).unwrap();
-        let mut sink = VideoSink::new(
-            i64::from(next_id()),
-            sys::VideoSinkInterface::create_forwarding(Box::new(
-                VideoFormatSink {
-                    width: Arc::clone(&width),
-                    height: Arc::clone(&height),
-                },
-            )),
-            VideoTrackId(track.id()),
-            track_origin,
-        );
-
-        let mut res = Self {
-            kind: kind::Video {
-                id: VideoTrackId(track.id()),
-                inner: track.try_into().unwrap(),
-                sinks: Vec::new(),
-                width,
-                height,
-                sink: None,
-                // Safe to unwrap since transceiver is guaranteed
-                // to be negotiated at this point.
-                source: MediaTrackSource::Remote {
-                    mid: transceiver.mid().unwrap(),
-                    peer: Arc::downgrade(peer),
-                },
-            },
-            senders: HashMap::new(),
-            events_tx: None,
-            origin: track_origin,
-        };
-
-        res.add_video_sink(&mut sink);
-        res.kind.sink = Some(sink);
-
-        res
-    }
-
-    /// Adds the provided [`VideoSink`] to this [`VideoTrack`].
-    pub fn add_video_sink(&mut self, video_sink: &mut VideoSink) {
-        self.kind.add_video_sink(video_sink);
-    }
-
-    /// Detaches the provided [`VideoSink`] from this [`VideoTrack`].
-    pub fn remove_video_sink(&mut self, video_sink: VideoSink) {
-        self.kind.remove_video_sink(video_sink);
-    }
-}
-
-impl From<&VideoTrack> for api::MediaStreamTrack {
-    fn from(track: &VideoTrack) -> Self {
-        Self {
-            id: track.id().0,
-            device_id: match track.source() {
-                MediaTrackSource::Local(src) => src.device_id.to_string(),
-                MediaTrackSource::Remote { .. } => "remote".into(),
-            },
-            kind: api::MediaType::Video,
-            enabled: true,
-            peer_id: match track.origin {
-                TrackOrigin::Local => None,
-                TrackOrigin::Remote(peer_id) => Some(peer_id.into()),
-            },
-        }
-    }
-}
-
-/// Representation of an `Audio` track interface.
-pub type AudioTrack = Track<kind::Audio>;
-
-impl AsRef<sys::AudioTrackInterface> for AudioTrack {
-    fn as_ref(&self) -> &sys::AudioTrackInterface {
-        self.kind.as_ref()
-    }
-}
-
-impl AudioTrack {
-    /// Returns ID of this [`AudioTrack`].
-    #[must_use]
-    pub fn id(&self) -> AudioTrackId {
-        self.kind.id.clone()
-    }
-
-    /// Returns [`AudioSource`] that is used by this [`AudioTrack`].
-    #[must_use]
-    pub const fn source(&self) -> &MediaTrackSource<AudioSource> {
-        &self.kind.source
-    }
-
-    /// Returns the [readyState][0] property of the
-    /// [`sys::AudioTrackInterface`].
-    ///
-    /// [0]: https://w3.org/TR/mediacapture-streams#dfn-readystate
-    fn state(&self) -> api::TrackState {
-        self.kind.inner.state().into()
-    }
-
-    /// Changes the [enabled][1] property of the [`sys::AudioTrackInterface`].
-    ///
-    /// [1]: https://w3.org/TR/mediacapture-streams#track-enabled
-    fn set_enabled(&self, enabled: bool) {
-        self.kind.inner.set_enabled(enabled);
-    }
-
-    /// Creates a new [`AudioTrack`].
-    ///
-    /// # Errors
-    ///
-    /// Whenever [`sys::PeerConnectionFactoryInterface::create_audio_track()`]
-    /// returns an error.
-    pub fn new(
-        pc: &sys::PeerConnectionFactoryInterface,
-        src: Arc<AudioSource>,
-        track_origin: TrackOrigin,
-    ) -> anyhow::Result<Self> {
-        let id = AudioTrackId(next_id().to_string());
-        Ok(Self {
-            kind: kind::Audio {
-                id: id.clone(),
-                inner: pc.create_audio_track(id.into(), &src.src)?,
-                volume_observer_id: None,
-                source: MediaTrackSource::Local(src),
-            },
-            senders: HashMap::new(),
-            origin: track_origin,
-            events_tx: None,
-        })
-    }
-
-    /// Subscribes this [`AudioTrack`] to audio level updates.
-    ///
-    /// Volume updates will be passed to the `stream_sink` of this
-    /// [`AudioTrack`].
-    pub fn subscribe_to_audio_level(&mut self) {
-        self.kind.subscribe_to_audio_level(self.events_tx.clone());
-    }
-
-    /// Unsubscribes this [`AudioTrack`] from audio level updates.
-    pub fn unsubscribe_from_audio_level(&self) {
-        self.kind.unsubscribe_from_audio_level();
-    }
-
-    /// Wraps the track of the `transceiver.receiver.track()` into an
-    /// [`AudioTrack`].
-    pub(crate) fn wrap_remote(
-        transceiver: &sys::RtpTransceiverInterface,
-        peer: &Arc<PeerConnection>,
-    ) -> Self {
-        let receiver = transceiver.receiver();
-        let track = receiver.track();
-        Self {
-            kind: kind::Audio {
-                id: AudioTrackId(track.id()),
-                inner: track.try_into().unwrap(),
-                volume_observer_id: None,
-                // Safe to unwrap since transceiver is guaranteed
-                // to be negotiated at this point.
-                source: MediaTrackSource::Remote {
-                    mid: transceiver.mid().unwrap(),
-                    peer: Arc::downgrade(peer),
-                },
-            },
-            senders: HashMap::new(),
-            origin: TrackOrigin::Remote(peer.id()),
-            events_tx: None,
-        }
-    }
-}
-impl From<&AudioTrack> for api::MediaStreamTrack {
-    fn from(track: &AudioTrack) -> Self {
-        Self {
-            id: track.id().0,
-            device_id: match track.source() {
-                MediaTrackSource::Local(local) => local.device_id.to_string(),
-                MediaTrackSource::Remote { mid: _, peer: _ } => "remote".into(),
-            },
-            kind: api::MediaType::Audio,
-            enabled: true,
-            peer_id: match track.origin {
-                TrackOrigin::Local => None,
-                TrackOrigin::Remote(peer_id) => Some(peer_id.into()),
-            },
-        }
-    }
-}
-
-/// [`sys::AudioSourceOnAudioLevelChangeCallback`] unique (per
-/// [`AudioSourceInterface`]) ID.
-#[derive(Clone, Copy, Debug, Default, Eq, Hash, PartialEq)]
-pub struct AudioLevelObserverId(u64);
-
-/// Storage for a [`sys::AudioSourceOnAudioLevelChangeCallback`].
-type ObserverStorage =
-    Arc<RwLock<HashMap<AudioLevelObserverId, AudioSourceAudioLevelHandler>>>;
-
-/// [`sys::AudioSourceOnAudioLevelChangeCallback`] implementation which
-/// broadcasts all audio level updates to all the underlying
-/// [`sys::AudioSourceOnAudioLevelChangeCallback`]s.
-struct BroadcasterObserver(ObserverStorage);
-
-impl BroadcasterObserver {
-    /// Creates a new [`BroadcasterObserver`] with the provided
-    /// [`ObserverStorage`] as a sink for audio level broadcasts.
-    pub const fn new(observers: ObserverStorage) -> Self {
-        Self(observers)
-    }
-}
-
-impl sys::AudioSourceOnAudioLevelChangeCallback for BroadcasterObserver {
-    /// Propagates audio level change to all the underlying
-    /// [`sys::AudioSourceOnAudioLevelChangeCallback`]s.
-    fn on_audio_level_change(&self, volume: f32) {
-        let observers = self.0.read().unwrap();
-
-        observers.values().for_each(|observer| {
-            observer.on_audio_level_change(volume);
-        });
-    }
-}
-
-/// [`sys::AudioSourceInterface`] wrapper.
-pub struct AudioSource {
-    /// Storage for the all the [`sys::AudioSourceOnAudioLevelChangeCallback`]
-    /// related to this [`AudioSourceInterface`].
-    ///
-    /// This [`ObserverStorage`] is shared with the [`BroadcasterObserver`] and
-    /// needed for a [`sys::AudioSourceOnAudioLevelChangeCallback`] disposing
-    /// without calling the C++ side.
-    observers: ObserverStorage,
-
-    /// Last ID used for the [`AudioLevelObserverId`].
-    last_observer_id: Mutex<AudioLevelObserverId>,
-
-    /// [`AudioDeviceId`] of the device this [`AudioSource`] is related to.
-    pub device_id: AudioDeviceId,
-
-    /// Underlying FFI wrapper for the `LocalAudioSource`.
-    src: sys::AudioSourceInterface,
-
-    /// [`sys::AudioProcessing`] used by this [`AudioSource`].
-    ap: sys::AudioProcessing,
-}
-
-impl AudioSource {
-    /// Creates a new [`AudioSource`] with the provided parameters.
-    #[must_use]
-    pub fn new(
-        device_id: AudioDeviceId,
-        src: sys::AudioSourceInterface,
-        ap: sys::AudioProcessing,
-    ) -> Self {
-        Self {
-            device_id,
-            observers: Arc::default(),
-            last_observer_id: Mutex::default(),
-            src,
-            ap,
-        }
-    }
-
-    /// Subscribes the provided [`sys::AudioSourceOnAudioLevelChangeCallback`]
-    /// to audio level updates.
-    ///
-    /// This method will initialize new [`BroadcasterObserver`] if it wasn't
-    /// initialized before.
-    ///
-    /// Returns [`AudioLevelObserverId`] which can be used to unsubscribe the
-    /// provided here [`sys::AudioSourceOnAudioLevelChangeCallback`].
-    ///
-    /// # Panics
-    ///
-    /// On [`Mutex`] poisoning.
-    fn subscribe_on_audio_level(
-        &self,
-        cb: AudioSourceAudioLevelHandler,
-    ) -> AudioLevelObserverId {
-        let mut observers = self.observers.write().unwrap();
-
-        if observers.is_empty() {
-            self.src.subscribe(Box::new(BroadcasterObserver::new(Arc::clone(
-                &self.observers,
-            ))));
-        }
-
-        let observer_id = {
-            let mut last_observer_id = self.last_observer_id.lock().unwrap();
-            let next_id = AudioLevelObserverId(last_observer_id.0 + 1);
-            *last_observer_id = next_id;
-            next_id
-        };
-        observers.insert(observer_id, cb);
-
-        observer_id
-    }
-
-    /// Unsubscribes the current [`sys::AudioSourceOnAudioLevelChangeCallback`]
-    /// from audio level updates.
-    ///
-    /// After unsubscribing this callback will be disposed.
-    ///
-    /// If [`AudioSourceInterface`] detects that this was the last callback, it
-    /// will stop any audio level calculations to save system resources.
-    ///
-    /// # Panics
-    ///
-    /// On [`Mutex`] poisoning.
-    fn unsubscribe_audio_level(&self, id: AudioLevelObserverId) {
-        let mut observers = self.observers.write().unwrap();
-
-        observers.remove(&id);
-        if observers.is_empty() {
-            self.src.unsubscribe();
-        }
-    }
-
-    /// Applies the provided [`api::AudioProcessingConstraints`] to the
-    /// [`sys::AudioProcessing`] of this [`AudioSource`].
-    fn update_audio_processing(
-        &self,
-        new_conf: &api::AudioProcessingConstraints,
-    ) {
-        let mut conf = self.ap.config();
-        if let Some(aec) = new_conf.echo_cancellation {
-            conf.set_echo_cancellation_enabled(aec);
-        }
-        if let Some(hpf) = new_conf.high_pass_filter {
-            conf.set_high_pass_filter_enabled(hpf);
-        }
-        if let Some(agc) = new_conf.auto_gain_control {
-            conf.set_gain_controller_enabled(agc);
-        }
-        if let Some(ns) = new_conf.noise_suppression {
-            conf.set_noise_suppression_enabled(ns);
-        }
-        if let Some(nsl) = new_conf.noise_suppression_level {
-            conf.set_noise_suppression_level(nsl.into());
-        }
-        self.ap.apply_config(&conf);
-    }
-}
-
-/// [`sys::VideoTrackSourceInterface`] wrapper.
-pub struct VideoSource {
-    /// Underlying [`sys::VideoTrackSourceInterface`].
-    inner: sys::VideoTrackSourceInterface,
-
-    /// ID of an video input device that provides data to this [`VideoSource`].
-    pub device_id: VideoDeviceId,
-}
-
-impl VideoSource {
-    /// Creates a new [`VideoTrackSourceInterface`] from the video input device
-    /// with the specified constraints.
-    fn new_device_source(
-        worker_thread: &mut sys::Thread,
-        signaling_thread: &mut sys::Thread,
-        caps: &api::VideoConstraints,
-        device_index: u32,
-        device_id: VideoDeviceId,
-    ) -> anyhow::Result<Self> {
-        let inner = if api::is_fake_media() {
-            sys::VideoTrackSourceInterface::create_fake(
-                worker_thread,
-                signaling_thread,
-                caps.width as usize,
-                caps.height as usize,
-                caps.frame_rate as usize,
-            )
-        } else {
-            sys::VideoTrackSourceInterface::create_proxy_from_device(
-                worker_thread,
-                signaling_thread,
-                caps.width as usize,
-                caps.height as usize,
-                caps.frame_rate as usize,
-                device_index,
-            )
-        }
-        .with_context(|| {
-            format!("Failed to acquire device with ID `{device_id}`")
-        })?;
-
-        Ok(Self { inner, device_id })
-    }
-
-    /// Starts screen capturing and creates a new [`VideoTrackSourceInterface`]
-    /// with the specified constraints.
-    fn new_display_source(
-        worker_thread: &mut sys::Thread,
-        signaling_thread: &mut sys::Thread,
-        caps: &api::VideoConstraints,
-        device_id: VideoDeviceId,
-    ) -> anyhow::Result<Self> {
-        let inner = if api::is_fake_media() {
-            sys::VideoTrackSourceInterface::create_fake(
-                worker_thread,
-                signaling_thread,
-                caps.width as usize,
-                caps.height as usize,
-                caps.frame_rate as usize,
-            )?
-        } else {
-            sys::VideoTrackSourceInterface::create_proxy_from_display(
-                worker_thread,
-                signaling_thread,
-                device_id.0.parse::<i64>()?,
-                caps.width as usize,
-                caps.height as usize,
-                caps.frame_rate as usize,
-            )?
-        };
-        Ok(Self { inner, device_id })
-    }
-}
-
-/// Wrapper around [`TrackObserverInterface`] implementing
-/// [`sys::TrackEventCallback`].
-struct TrackEventHandler(StreamSink<api::TrackEvent>);
-
-impl TrackEventHandler {
-    /// Creates a new [`TrackEventHandler`] with the provided [`StreamSink`].
-    ///
-    /// Sends an [`api::TrackEvent::TrackCreated`] to the provided
-    /// [`StreamSink`].
-    pub fn new(cb: StreamSink<api::TrackEvent>) -> Self {
-        _ = cb.add(api::TrackEvent::TrackCreated);
-        Self(cb)
-    }
-}
-
-impl sys::TrackEventCallback for TrackEventHandler {
-    fn on_ended(&mut self) {
-        _ = self.0.add(api::TrackEvent::Ended);
-    }
-}
-
-/// Wrapper around a [`StreamSink`] which emits [`AudioLevelUpdated`] events to
-/// the Flutter side.
-///
-/// This handler also will multiply volume by `1000` and cast it to [`u32`], for
-/// more convenient usage.
-///
-/// [`AudioLevelUpdated`]: api::TrackEvent::AudioLevelUpdated
-struct AudioSourceAudioLevelHandler(StreamSink<api::TrackEvent>);
-
-impl AudioSourceAudioLevelHandler {
-    #[expect( // intentional
-        clippy::cast_possible_truncation,
-        clippy::cast_sign_loss,
-        reason = "size fits and non-negative"
-    )]
-    fn on_audio_level_change(&self, level: f32) {
-        _ = self.0.add(api::TrackEvent::AudioLevelUpdated(cmp::min(
-            (level * 1000.0).round() as u32,
-            100,
-        )));
-    }
-}
-
-impl From<&api::AudioProcessingConstraints> for sys::AudioProcessingConfig {
-    fn from(caps: &api::AudioProcessingConstraints) -> Self {
-        let mut conf = Self::default();
-
-        conf.set_high_pass_filter_enabled(
-            caps.high_pass_filter.unwrap_or(true),
-        );
-        conf.set_echo_cancellation_enabled(
-            caps.echo_cancellation.unwrap_or(true),
-        );
-        conf.set_gain_controller_enabled(
-            caps.auto_gain_control.unwrap_or(true),
-        );
-
-        conf.set_noise_suppression_enabled(
-            caps.noise_suppression.unwrap_or(true),
-        );
-        conf.set_noise_suppression_level(
-            caps.noise_suppression_level
-                .unwrap_or(api::NoiseSuppressionLevel::VeryHigh)
-                .into(),
-        );
-
-        conf
-    }
-}
